requirements.txt
Flask
gunicorn
Flask-Login
Flask-WTF
email-validator
Flask-SQLAlchemy
PyMySQL
python-dotenv
